modules: location: channel definition into module header
Location channel definition into dedicated location header.
