+++ app/src/common/message_channel.c
@@ -49,11 +49,3 @@ ZBUS_CHAN_DEFINE(TIME_CHAN,
 		 ZBUS_OBSERVERS_EMPTY,
 		 ZBUS_MSG_INIT(0)
 );
-
-ZBUS_CHAN_DEFINE(LOCATION_CHAN,
-		 enum location_status,
-		 NULL,
-		 NULL,
-		 ZBUS_OBSERVERS_EMPTY,
-		 ZBUS_MSG_INIT(0)
-);
