+++ app/src/modules/location/CMakeLists.txt
@@ -5,3 +5,4 @@
 #
 
 target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/location.c)
+target_include_directories(app PRIVATE .)
